gltransformation: implement pivot point for rotation and scale
https://bugzilla.gnome.org/show_bug.cgi?id=744763

Add an pivot vector for setting the origin of rotations and scales.

With the pivot point the rotation and scale operations can have
different origins. This adds the ability to rotate around different points.
Currently the default (0, 0) pivot point is possible. Roation around the center,
also scaling / zooming into the center.

With an pivot point this is optional.
I defined the following image coordinates for the pivot point:

	 (-1,1) ----------------- (1,1)
		|		|
		|		|
		|	(0,0)	|
		|		|
		|		|
		|		|
	(-1,-1) ----------------- (1,-1)

Example:

	Rotate the video at the bottom left corner

	gst-launch-1.0 videotestsrc \
			! gltransformation \
		 		scale-x=0.5 \
				scale-y=0.5 \
				rotation-z=25.0 \
				pivot-x=-1.0 \
				pivot-y=-1.0 \
			! glimagesink

The pivot-z option defined the pivot point in 3D space,
this only affects the rotation, since we have no Z data to scale.
With this option a video can be rotated around a point in 3D space.